IPRT/tstRTInlineAsm: Fix harmless parfait warning. bugref:3409
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@103354 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/testcase/tstRTInlineAsm.cpp
@@ -2452,8 +2452,8 @@ static void tstASMMemFirstMismatchingU8(RTTEST hTest)
     size_t const   cbBuf    = 128;
     uint8_t       *pbBuf1   = pbPage1;
     uint8_t       *pbBuf2   = &pbPage2[PAGE_SIZE - cbBuf]; /* Put it up against the tail guard */
-    memset(pbPage1, ~bFiller1, PAGE_SIZE);
-    memset(pbPage2, ~bFiller2, PAGE_SIZE);
+    memset(pbPage1, (uint8_t)~bFiller1, PAGE_SIZE);
+    memset(pbPage2, (uint8_t)~bFiller2, PAGE_SIZE);
     memset(pbBuf1, bFiller1, cbBuf);
     memset(pbBuf2, bFiller2, cbBuf);
     for (size_t offNonZero = 0; offNonZero < cbBuf; offNonZero++)
